+++ sdk/include/opentelemetry/sdk/logs/simple_log_record_processor.h
@@ -5,6 +5,7 @@
 #ifdef ENABLE_LOGS_PREVIEW
 
 #  include <atomic>
+#  include <memory>
 #  include <mutex>
 
 #  include "opentelemetry/common/spin_lock_mutex.h"